split transport manager out
