+++ rings-core/src/tests/wasm/test_wasm_transport.rs
@@ -16,7 +16,7 @@ use crate::prelude::RTCSdpType;
 use crate::session::SessionManager;
 use crate::storage::PersistenceStorage;
 use crate::swarm::Swarm;
-use crate::swarm::TransportManager;
+use crate::transports::manager::TransportManager;
 use crate::transports::Transport;
 use crate::types::channel::Channel;
 use crate::types::channel::Event;
