+++ src/browser/mod.rs
@@ -32,8 +32,8 @@ use crate::prelude::rings_core::session::SessionManager;
 use crate::prelude::rings_core::session::Signer;
 use crate::prelude::rings_core::storage::PersistenceStorage;
 use crate::prelude::rings_core::swarm::Swarm;
-use crate::prelude::rings_core::swarm::TransportManager;
 use crate::prelude::rings_core::transports::Transport;
+use crate::prelude::rings_core::transports::TransportManager;
 use crate::prelude::rings_core::types::ice_transport::IceTransportInterface;
 use crate::prelude::rings_core::types::message::MessageListener;
 use crate::prelude::wasm_bindgen;
